Move to pyproject.toml; CI python 3.12.
