+++ .github/workflows/build.yml
@@ -12,11 +12,11 @@ jobs:
       matrix:
         include:
         - {os: ubuntu-latest, python-version: "3.7"}
-        - {os: ubuntu-latest, python-version: "3.11"}
+        - {os: ubuntu-latest, python-version: "3.12"}
         - {os: macos-latest, python-version: "3.9"}
-        - {os: macos-latest, python-version: "3.11"}
+        - {os: macos-latest, python-version: "3.12"}
         - {os: windows-latest, python-version: "3.6"}
-        - {os: windows-latest, python-version: "3.11"}
+        - {os: windows-latest, python-version: "3.12"}
       fail-fast: false
     runs-on: ${{ matrix.os }}
     steps:
